pdf_bot/gemini_gem_docker/requirements.txt
psycopg2-binary
requests
duckduckgo-search
schedule
beautifulsoup4
